requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
args==0.1.0
async-timeout==5.0.1
atom==0.10.5
attrs==25.2.0
azure-core==1.32.0
azure-storage-blob==12.25.0
b2sdk==2.8.0
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.37.11
botocore==1.37.11
boxsdk==3.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clint==0.5.1
cryptography==43.0.3
debtcollector==3.0.0
deprecation==2.1.0
dropbox==12.0.2
duplicity==3.0.4.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fasteners==0.19
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
gdata-python3==3.0.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
gotrue==2.11.4
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ip_associations_python_novaclient_ext==0.2
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jottalib==0.5.1
keyring==25.6.0
keystoneauth1==2.18.0
logfury==1.0.1
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
mediafire==0.6.1
megatools==0.0.4
mock==5.2.0
monotonic==1.6
more-itertools==10.6.0
msgpack-python==0.5.6
multidict==6.1.0
netaddr==1.3.0
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.2
os_diskconfig_python_novaclient_ext==0.1.3
os_networksv2_python_novaclient_ext==0.26
os_virtual_interfacesv2_python_novaclient_ext==0.20
oslo.config==4.12.0
oslo.i18n==3.12.0
oslo.serialization==2.16.1
oslo.utils==3.22.3
packaging==24.2
paramiko==3.5.1
pbr==1.10.0
pexpect==4.9.0
ply==3.11
positional==1.2.1
postgrest==0.19.3
prettytable==0.7.2
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDrive2==1.21.3
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyrax==1.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gettext==5.0
python-keystoneclient==3.10.0
python-novaclient==2.27.0
python-swiftclient==4.7.0
pytz==2025.1
PyYAML==6.0.2
rackspace-auth-openstack==1.3
rackspace-novaclient==2.1
rax_default_network_flags_python_novaclient_ext==0.4.0
rax_scheduled_images_python_novaclient_ext==0.3.1
realtime==2.4.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rsa==4.9
s3transfer==0.11.4
SecretStorage==3.3.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stevedore==1.20.1
stone==3.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tlslite-ng==0.8.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
